+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -29,6 +30,7 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        //Gate::authorize('is-my-post', $request);
         $posts = Post::create([
             "user_id"=>Auth::user()->id,
             "title"=>$request->title,
@@ -49,6 +51,13 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        Gate::authorize('is-my-post', $post);
+
+        if (!$post) {
+            return response()->json([
+                'message' => 'Post not found'
+            ]);
+        }
         return $post;
     }
 
@@ -61,15 +70,23 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        $user_id=$request->user_id;
-        $title=$request->fulltitle;
+        //$user_id=$request->user_id;
+        $title=$request->title;
         $description=$request->description;
 
+        Gate::authorize('is-my-post', $post);
+
         $post = $post->update([
-            "user_id"=>$user_id,
+            "user_id"=>Auth::user()->id,
             "title"=>$title,
             "description"=>$description,
         ]);
+
+        if (!$post) {
+            return response()->json([
+                'message' => 'Post not found'
+            ]);
+        }
 
         return response()->json([
             'message' => 'Succesful',
@@ -85,6 +102,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('is-my-post', $post);
+
+        if (!$post) {
+            return response()->json([
+                'message' => 'Post not found'
+            ]);
+        }
         $post->delete();
     }
 }
